Color to float + ButtonGrid cleaning

+++ WaveFunctionCollapse/Color.h
@@ -3,12 +3,12 @@
 struct Color {
 public:
 	Color() : r(0), g(0), b(0), a(0) {};
-	Color(const unsigned int r, const unsigned int g, const unsigned int b, const unsigned int a) : r(r), g(g), b(b), a(a) {}
+	Color(const float r, const float g, const float b, const float a) : r(r), g(g), b(b), a(a) {}
 	~Color() {}
 
 public:
-	unsigned int r;
-	unsigned int g;
-	unsigned int b;
-	unsigned int a;
+	float r;
+	float g;
+	float b;
+	float a;
 };
